refactored netifaces, added *.pyc to gitignore
