+++ hunter.py
@@ -7,27 +7,28 @@ Purpose: To scan a network for a smb ports and validate if credentials work on t
 
 Copyright (c) 2015, Christopher Duffy All rights reserved.
 
-Redistribution and use in source and binary forms, with or without modification, 
-are permitted provided that the following conditions are met: * Redistributions 
-of source code must retain the above copyright notice, this list of conditions and 
-the following disclaimer. * Redistributions in binary form must reproduce the above 
-copyright notice, this list of conditions and the following disclaimer in the 
-documentation and/or other materials provided with the distribution. * Neither the 
-name of the nor the names of its contributors may be used to endorse or promote 
+Redistribution and use in source and binary forms, with or without modification,
+are permitted provided that the following conditions are met: * Redistributions
+of source code must retain the above copyright notice, this list of conditions and
+the following disclaimer. * Redistributions in binary form must reproduce the above
+copyright notice, this list of conditions and the following disclaimer in the
+documentation and/or other materials provided with the distribution. * Neither the
+name of the nor the names of its contributors may be used to endorse or promote
 products derived from this software without specific prior written permission.
 
-THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS 
-"AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, 
-THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE 
-DISCLAIMED. IN NO EVENT SHALL CHRISTOPHER DUFFY BE LIABLE FOR ANY DIRECT, INDIRECT, 
-INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, 
-PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS 
-INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT 
-LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE 
+THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+"AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO,
+THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+DISCLAIMED. IN NO EVENT SHALL CHRISTOPHER DUFFY BE LIABLE FOR ANY DIRECT, INDIRECT,
+INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
+PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
+LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 '''
 
 import os, argparse, sys, time
+import utilities.network as network
 try:
     import msfrpc
 except:
@@ -36,48 +37,6 @@ try:
     import nmap
 except:
     sys.exit("[!] Install the nmap library: pip install python-nmap")
-try:
-    import netifaces
-except:
-    sys.exit("[!] Install the netifaces library: pip install netifaces")
-
-def get_interfaces():
-    interfaces = netifaces.interfaces()
-    return interfaces
-
-def get_gateways():
-    gateway_dict = {}
-    gws = netifaces.gateways()
-    for gw in gws:
-        try:
-            gateway_iface = gws[gw][netifaces.AF_INET]
-            gateway_ip, iface = gateway_iface[0], gateway_iface[1]
-            gw_list =[gateway_ip, iface]
-            gateway_dict[gw]=gw_list
-        except:
-            pass
-    return gateway_dict
-
-def get_addresses(interface):
-    addrs = netifaces.ifaddresses(interface)
-    link_addr = addrs[netifaces.AF_LINK]
-    iface_addrs = addrs[netifaces.AF_INET]
-    iface_dict = iface_addrs[0]
-    link_dict = link_addr[0]
-    hwaddr = link_dict.get('addr')
-    iface_addr = iface_dict.get('addr')
-    iface_broadcast = iface_dict.get('broadcast')
-    iface_netmask = iface_dict.get('netmask')
-    return hwaddr, iface_addr, iface_broadcast, iface_netmask
-
-def get_networks(gateways_dict):
-    networks_dict = {}
-    for key, value in gateways_dict.iteritems():
-        gateway_ip, iface = value[0], value[1]
-        hwaddress, addr, broadcast, netmask = get_addresses(iface)
-        network = {'gateway': gateway_ip, 'hwaddr' : hwaddress, 'addr' : addr, 'broadcast' : broadcast, 'netmask' : netmask}
-        networks_dict[iface] = network
-    return networks_dict
 
 def target_identifier(verbose, dir, user, passwd, ips, port_num, ifaces, ipfile):
     hostlist = []
@@ -233,8 +192,8 @@ def main():
                 if verbose > 1:
                     print("[*] Using filename: %s") % (filename)
 
-    gateways = get_gateways()
-    network_ifaces = get_networks(gateways)
+    gateways = network.get_gateways()
+    network_ifaces = network.get_networks(gateways)
     hosts_file, hostlist = target_identifier(verbose, home_dir, username, password, targets, ports, network_ifaces, targets_file)
     run_commands(verbose, hostlist, username, password, domain, ports, filename)
